Extract GeneratorFunctionBase to eliminate code duplication
- Create GeneratorFunctionBase abstract class with shared infrastructure:
  - Property management (name, length, prototype, call/apply/bind)
  - Private slots handling
  - Extensibility control and interface implementations
  - ~500 lines of code now shared instead of duplicated

- Refactor TypedGeneratorFactory (function*) to extend base:
  - Only implements Invoke() and EnsureIntrinsics()
  - Reduced from ~630 to ~220 lines

- Refactor AsyncGeneratorFactory (async function*) to extend base:
  - Only implements Invoke(), EnsureIntrinsics(), CustomizePrototypeObject()
  - Reduced from ~615 to ~240 lines

This addresses Phase 2a of the ExecutionPlan refactoring plan (Option 4).
The thin wrapper pattern is now applied at the Factory layer, matching
the existing Instance layer pattern with ExecutionPlanRunner.
