+++ src/Asynkron.JsEngine/Ast/TypedAstEvaluator.AsyncGeneratorFactory.cs
@@ -1,7 +1,5 @@
 #region
 
-using System.Collections.Immutable;
-using System.Globalization;
 using Asynkron.JsEngine.JsTypes;
 using Asynkron.JsEngine.Parser;
 using Asynkron.JsEngine.Runtime;
@@ -13,19 +11,12 @@ namespace Asynkron.JsEngine.Ast;
 
 public static partial class TypedAstEvaluator
 {
-    private sealed class AsyncGeneratorFactory : IJsCallable, IJsObjectLike, IPropertyDefinitionHost,
-        IExtensibilityControl,
-        IFunctionNameTarget, ICallableMetadata
+    /// <summary>
+    /// Callable for async generator functions (async function*).
+    /// Returns an async iterator when invoked.
+    /// </summary>
+    private sealed class AsyncGeneratorFactory : GeneratorFunctionBase
     {
-        private readonly JsEnvironment _closure;
-        private readonly FunctionExpression _function;
-        private readonly bool _hasFunctionNameEnvironment;
-        private readonly bool _isLexicallyStrict;
-        private readonly JsObject _properties = new();
-        private ImmutableArray<PrivateNameScope> _capturedPrivateNameScopes = ImmutableArray<PrivateNameScope>.Empty;
-        private IJsObjectLike? _homeObject;
-        private bool _isConstructorEnabled;
-
         public AsyncGeneratorFactory(
             FunctionExpression function,
             JsEnvironment closure,
@@ -33,80 +24,19 @@ public static partial class TypedAstEvaluator
             bool isLexicallyStrict,
             bool hasFunctionNameEnvironment = false,
             bool isConstructorFunction = true)
+            : base(function, closure, realmState, isLexicallyStrict, hasFunctionNameEnvironment, isConstructorFunction)
         {
             if (!function.IsGenerator || !function.IsAsync)
             {
                 throw new ArgumentException("Factory can only wrap async generator functions.", nameof(function));
             }
 
-            _function = function;
-            _closure = closure;
-            RealmState = realmState;
-            _isLexicallyStrict = isLexicallyStrict;
-            _hasFunctionNameEnvironment = hasFunctionNameEnvironment;
-            _isConstructorEnabled = isConstructorFunction;
             InitializeProperties();
         }
 
-        public PrivateNameScope? PrivateNameScope { get; private set; }
+        protected override string FunctionTypeName => "AsyncGeneratorFunction";
 
-        public bool IsArrowFunction => false;
-        public bool DisallowConstruct => true;
-        public RealmState RealmState { get; }
-
-        public bool IsExtensible => _properties.IsExtensible;
-
-        public void PreventExtensions()
-        {
-            _properties.PreventExtensions();
-        }
-
-        public void EnsureHasName(string name, bool overwriteExisting = false)
-        {
-            if (string.IsNullOrEmpty(name))
-            {
-                return;
-            }
-
-            if (!overwriteExisting && _function.Name is not null)
-            {
-                return;
-            }
-
-            var descriptor = _properties.GetOwnPropertyDescriptor("name");
-            if (descriptor is { Configurable: false })
-            {
-                return;
-            }
-
-            if (!overwriteExisting && descriptor is not null)
-            {
-                if (descriptor.IsAccessorDescriptor || descriptor.JsValue.TryGetObject<IJsCallable>(out _))
-                {
-                    return;
-                }
-
-                if (descriptor.JsValue.TryGetString(out var existingName) && existingName.Length > 0)
-                {
-                    return;
-                }
-            }
-
-            _properties.DefineProperty("name",
-                new PropertyDescriptor
-                {
-                    JsValue = new JsValue(name),
-                    Writable = false,
-                    Enumerable = false,
-                    Configurable = true,
-                    HasValue = true,
-                    HasWritable = true,
-                    HasEnumerable = true,
-                    HasConfigurable = true
-                });
-        }
-
-        public JsValue Invoke(IReadOnlyList<JsValue> arguments, JsValue thisValue)
+        public override JsValue Invoke(IReadOnlyList<JsValue> arguments, JsValue thisValue)
         {
             var instance = new AsyncGeneratorInstance(
                 _function,
@@ -124,200 +54,7 @@ public static partial class TypedAstEvaluator
             return (JsValue)instance.CreateAsyncIteratorObject();
         }
 
-        public JsObject? Prototype => _properties.Prototype;
-
-        public bool IsSealed => _properties.IsSealed;
-        public bool IsFrozen => _properties.IsFrozen;
-
-        public IEnumerable<string> Keys => _properties.Keys;
-
-        public void DefineProperty(string name, PropertyDescriptor descriptor)
-        {
-            _properties.DefineProperty(name, descriptor);
-        }
-
-        public void SetPrototype(IJsPropertyAccessor? candidate)
-        {
-            _properties.SetPrototype(candidate);
-        }
-
-        public void Seal()
-        {
-            _properties.Seal();
-        }
-
-        public bool Delete(string name)
-        {
-            return _properties.DeleteOwnProperty(name);
-        }
-
-        public bool TryGetProperty(string name, JsValue receiver, out JsValue value)
-        {
-            // Handle call/apply/bind specially BEFORE looking them up in prototype chain
-            // This ensures async generator functions get proper constructor semantics for bound functions
-            var callable = (IJsCallable)this;
-            switch (name)
-            {
-                case "call":
-                    value = (JsValue)new HostFunction((_, args) =>
-                    {
-                        var thisArg = args.GetArgument(0);
-                        var callArgs = args.SliceFrom(1);
-                        return callable.Invoke(callArgs, thisArg);
-                    });
-                    return true;
-
-                case "apply":
-                    value = (JsValue)new HostFunction((_, args) =>
-                    {
-                        var thisArg = args.GetArgument(0);
-                        IReadOnlyList<JsValue> argList = ArgumentSlice.Empty;
-                        if (args.Count > 1 && args[1].TryUnwrap(out JsArray? jsArray))
-                        {
-                            var items = jsArray.Items;
-                            var converted = new JsValue[items.Count];
-                            for (var i = 0; i < items.Count; i++)
-                            {
-                                // items[i] is already JsValue from JsArray.Items
-                                converted[i] = items[i];
-                            }
-
-                            argList = converted;
-                        }
-
-                        return callable.Invoke(argList, thisArg);
-                    });
-                    return true;
-
-                case "bind":
-                    value = (JsValue)new HostFunction((_, args) =>
-                    {
-                        var boundThis = args.GetArgument(0);
-                        var boundArgs = args.SliceFrom(1);
-
-                        // Async generator functions are never constructors, so bound async generator functions
-                        // must also have DisallowConstruct = true per ES spec.
-                        return (JsValue)new HostFunction((_, innerArgs) =>
-                        {
-                            if (boundArgs.Count == 0)
-                            {
-                                return callable.Invoke(innerArgs, boundThis);
-                            }
-
-                            if (innerArgs.Count == 0)
-                            {
-                                return callable.Invoke(boundArgs, boundThis);
-                            }
-
-                            var finalArgs = new JsValue[boundArgs.Count + innerArgs.Count];
-                            for (var i = 0; i < boundArgs.Count; i++)
-                            {
-                                finalArgs[i] = boundArgs[i];
-                            }
-
-                            for (var i = 0; i < innerArgs.Count; i++)
-                            {
-                                finalArgs[boundArgs.Count + i] = innerArgs[i];
-                            }
-
-                            return callable.Invoke(finalArgs, boundThis);
-                        }, RealmState, false) { DisallowConstruct = true };
-                    });
-                    return true;
-            }
-
-            // Fall back to properties lookup for all other properties
-            var receiverValue = receiver.IsUndefined ? JsValue.FromObjectUnsafe(this) : receiver;
-            if (_properties.TryGetProperty(name, receiverValue, out var objValue))
-            {
-                value = objValue;
-                return true;
-            }
-
-            value = JsValue.Undefined;
-            return false;
-        }
-
-        public bool TryGetProperty(string name, out JsValue value)
-        {
-            return TryGetProperty(name, JsValue.FromObjectUnsafe(this), out value);
-        }
-
-
-        public void SetProperty(string name, JsValue value)
-        {
-            SetProperty(name, value, JsValue.FromObjectUnsafe(this));
-        }
-
-
-        public void SetProperty(string name, JsValue value, JsValue receiver)
-        {
-            var receiverValue = receiver.IsUndefined ? JsValue.FromObjectUnsafe(this) : receiver;
-            _properties.SetProperty(name, value, receiverValue);
-        }
-
-        PropertyDescriptor? IJsPropertyAccessor.GetOwnPropertyDescriptor(string name)
-        {
-            var descriptor = _properties.GetOwnPropertyDescriptor(name);
-            if (descriptor is not null && string.Equals(name, "name", StringComparison.Ordinal))
-            {
-                descriptor.Writable = false;
-                descriptor.Enumerable = false;
-                descriptor.Configurable = true;
-            }
-
-            return descriptor;
-        }
-
-        public IEnumerable<string> GetOwnPropertyNames()
-        {
-            return _properties.GetOwnPropertyNames();
-        }
-
-        public IEnumerable<string> GetEnumerablePropertyNames()
-        {
-            return _properties.GetEnumerablePropertyNames();
-        }
-
-        public bool TryDefineProperty(string name, PropertyDescriptor descriptor)
-        {
-            return _properties.TryDefineProperty(name, descriptor);
-        }
-
-        public void SetPrivateNameScope(PrivateNameScope? scope)
-        {
-            PrivateNameScope = scope;
-        }
-
-        public void SetCapturedPrivateNameScopes(ImmutableArray<PrivateNameScope> scopes)
-        {
-            _capturedPrivateNameScopes = scopes;
-        }
-
-        public void DisableConstruction()
-        {
-            if (!_isConstructorEnabled)
-            {
-                return;
-            }
-
-            _isConstructorEnabled = false;
-            _properties.DeleteOwnProperty("prototype");
-        }
-
-        public void SetHomeObject(IJsObjectLike homeObject)
-        {
-            _homeObject = homeObject;
-        }
-
-        public override string ToString()
-        {
-            return _function.Name is { } name
-                ? $"[AsyncGeneratorFunction: {name.Name}]"
-                : "[AsyncGeneratorFunction]";
-        }
-
-        private void EnsureAsyncGeneratorIntrinsics()
+        protected override void EnsureIntrinsics()
         {
             var engine = RealmState.Engine ?? throw new InvalidOperationException("Engine reference is missing.");
 
@@ -386,6 +123,33 @@ public static partial class TypedAstEvaluator
 
                 RealmState.AsyncGeneratorFunctionPrototype = asyncGenFuncProto;
             }
+        }
+
+        protected override IJsPropertyAccessor? GetFunctionPrototype()
+        {
+            return RealmState.AsyncGeneratorFunctionPrototype;
+        }
+
+        protected override JsObject? GetGeneratorPrototype()
+        {
+            return RealmState.AsyncGeneratorPrototype ?? RealmState.ObjectPrototype;
+        }
+
+        protected override void CustomizePrototypeObject(JsObject prototypeObject)
+        {
+            // Async generators add a constructor property pointing to themselves
+            prototypeObject.DefineProperty("constructor",
+                new PropertyDescriptor
+                {
+                    Value = this,
+                    Writable = true,
+                    Enumerable = false,
+                    Configurable = true,
+                    HasValue = true,
+                    HasWritable = true,
+                    HasEnumerable = true,
+                    HasConfigurable = true
+                });
         }
 
         private static HostFunction CreateAsyncGeneratorFunctionConstructor(JsEngine engine, RealmState realm)
@@ -472,143 +236,6 @@ public static partial class TypedAstEvaluator
             }
 
             return created;
-        }
-
-        private static string ToFunctionArgumentString(JsValue value, EvaluationContext evalContext, RealmState realm)
-        {
-            var primitiveObj = JsOps.ToPrimitive(value, ToPrimitiveHint.String, evalContext);
-            if (evalContext.IsThrow)
-            {
-                throw new ThrowSignal(evalContext.FlowValue);
-            }
-
-            var primitive = primitiveObj;
-
-            if (primitive.IsNull)
-            {
-                return "null";
-            }
-
-            if (primitive.IsUndefined)
-            {
-                return "undefined";
-            }
-
-            if (primitive.TryUnwrap(out Symbol? _) || primitive.TryUnwrap(out TypedAstSymbol? _))
-            {
-                throw StandardLibrary.ThrowTypeError("Cannot convert a Symbol value to a string", evalContext, realm);
-            }
-
-            if (primitive.TryGetBoolean(out var flag))
-            {
-                return flag ? "true" : "false";
-            }
-
-            if (primitive.TryGetString(out var s))
-            {
-                return s ?? string.Empty;
-            }
-
-            if (primitive.TryUnwrap(out JsBigInt? bigInt))
-            {
-                return bigInt.Value.ToString(CultureInfo.InvariantCulture);
-            }
-
-            if (primitive.TryGetDouble(out var d))
-            {
-                if (double.IsNaN(d))
-                {
-                    return "NaN";
-                }
-
-                if (double.IsPositiveInfinity(d))
-                {
-                    return "Infinity";
-                }
-
-                if (double.IsNegativeInfinity(d))
-                {
-                    return "-Infinity";
-                }
-
-                return d.ToString(CultureInfo.InvariantCulture);
-            }
-
-            // At this point, all primitive types have been handled above;
-            // remaining cases are objects, so use ObjectValue directly.
-            return Convert.ToString(primitive.ObjectValue, CultureInfo.InvariantCulture) ?? string.Empty;
-        }
-
-        private void InitializeProperties()
-        {
-            EnsureAsyncGeneratorIntrinsics();
-
-            if (RealmState.AsyncGeneratorFunctionPrototype is { } asyncGenFuncProto)
-            {
-                _properties.SetPrototype(asyncGenFuncProto);
-            }
-            else if (RealmState.FunctionPrototype is { } functionPrototype)
-            {
-                _properties.SetPrototype(functionPrototype);
-            }
-
-            if (_isConstructorEnabled && RealmState.ObjectPrototype is not null)
-            {
-                var generatorPrototype = new JsObject();
-                generatorPrototype.SetPrototype(RealmState.AsyncGeneratorPrototype ?? RealmState.ObjectPrototype);
-                generatorPrototype.DefineProperty("constructor",
-                    new PropertyDescriptor
-                    {
-                        Value = this,
-                        Writable = true,
-                        Enumerable = false,
-                        Configurable = true,
-                        HasValue = true,
-                        HasWritable = true,
-                        HasEnumerable = true,
-                        HasConfigurable = true
-                    });
-                _properties.DefineProperty("prototype",
-                    new PropertyDescriptor
-                    {
-                        Value = generatorPrototype,
-                        Writable = true,
-                        Enumerable = false,
-                        Configurable = false,
-                        HasValue = true,
-                        HasWritable = true,
-                        HasEnumerable = true,
-                        HasConfigurable = true
-                    });
-            }
-
-            var paramCount = _function.Parameters.GetExpectedParameterCount();
-            _properties.DefineProperty("length",
-                new PropertyDescriptor
-                {
-                    Value = (double)paramCount,
-                    Writable = false,
-                    Enumerable = false,
-                    Configurable = true,
-                    HasValue = true,
-                    HasWritable = true,
-                    HasEnumerable = true,
-                    HasConfigurable = true
-                });
-
-            var functionNameValue = _function.Name?.Name ?? string.Empty;
-            _properties.DefineProperty("name",
-                new PropertyDescriptor
-                {
-                    Value = functionNameValue,
-                    Writable = false,
-                    Enumerable = false,
-                    Configurable = true,
-                    HasValue = true,
-                    HasWritable = true,
-                    HasEnumerable = true,
-                    HasConfigurable = true
-                });
         }
     }
 }
